code made modular on portfolio screen.
